requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
WTForms==3.0.1
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
plotly==5.16.1
passlib==1.7.4
email-validator==2.0.0.post2
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
SQLAlchemy==2.0.20
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
requests==2.31.0
psycopg2-binary==2.9.9
